Refactors auth response handling and logging in pubkey hook handler; introduces `AuthResponse` model; consolidates logging into `authHookEvents`
