+++ internal/handlers/models/auth_response.go
@@ -1,14 +1,6 @@
 package models
 
-type PubKeyRequest struct {
-	ClientVersion string `json:"clientVersion"`
-	ConnectionId  string `json:"connectionId"`
-	RemoteAddress string `json:"remoteAddress"`
-	Username      string `json:"username"`
-	PublicKey     string `json:"publicKey"`
-}
-
-type PubKeyResponse struct {
+type AuthResponse struct {
 	AuthenticatedUsername string `json:"authenticatedUsername"`
 	ClientVersion         string `json:"clientVersion"`
 	ConnectionId          string `json:"connectionId"`
