+++ internal/services/containerssh_docker_appconfig_service_test.go
@@ -2,12 +2,13 @@ package services
 
 import (
 	"errors"
+	"testing"
+
 	"github.com/matzefriedrich/containerssh-authserver/internal/configuration"
 	"github.com/matzefriedrich/containerssh-authserver/internal/shims"
 	"github.com/matzefriedrich/parsley/pkg/features"
 	"github.com/rs/zerolog"
 	"github.com/stretchr/testify/assert"
-	"testing"
 )
 
 func Test_CreateApplicationConfigFor_returns_error_if_user_not_found(t *testing.T) {
@@ -53,7 +54,7 @@ func Test_CreateApplicationConfigFor_returns_docker_container_configuration_for_
 
 	// Assert
 	assert.NoError(t, err)
-	assert.Equal(t, expectedDockerImage, actual.Docker.Execution.DockerLaunchConfigShim.ContainerConfig.Image)
+	assert.Equal(t, expectedDockerImage, actual.Docker.Execution.ContainerConfig.Image)
 
 	_, found := actual.Docker.Execution.NetworkConfig.EndpointsConfig[expectedContainerNetworkName]
 	assert.True(t, found)
